add: indices must use Uint16Array

+++ main.js
@@ -121,8 +121,8 @@ function main() {
     // Create a linked-list for storing the vertices data
     var index_buffer = gl.createBuffer();
     gl.bindBuffer(gl.ARRAY_BUFFER, index_buffer);
-    gl.bufferData(gl.ARRAY_BUFFER, new Float32Array(indices), gl.STATIC_DRAW);
-    
+    gl.bufferData(gl.ARRAY_BUFFER, new Uint16Array(indices), gl.STATIC_DRAW);
+
 
     var vertexShaderSource = `
         attribute vec2 aPosition;
